+++ aclients/aio_http_client.py
@@ -12,10 +12,9 @@ import atexit
 import aelog
 import aiohttp
 
-from .decorators import Singleton
 from .err_msg import http_msg
 from .exceptions import ClientConnectionError, ClientError, ClientResponseError, HttpError
-from .utils import verify_message
+from .utils import Singleton, verify_message
 
 __all__ = ("AIOHttpClient", "AsyncResponse")